chore: move QuotesView utils fn to own file
